pc_donation/requirements.txt
alembic==1.7.1
azure-ai-formrecognizer==3.1.2
azure-ai-textanalytics==5.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-contentmoderator==1.0.0
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.27
azure-core==1.18.0
azure-identity==1.6.1
azure-keyvault-secrets==4.3.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-common==2.1.0
Babel==2.9.1
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cognitive-face==1.5.0
colorclass==2.2.0
cryptography==3.4.8
dnspython==2.1.0
docopt==0.6.2
dominate==2.6.0
email-validator==1.1.3
filelock==3.0.12
Flask==1.1.4
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Classful==0.14.2
Flask-Datepicker==0.14
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
ghp-import==2.0.1
google-api-core==1.26.3
google-auth==1.30.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
htmlmin==0.1.12
idna==3.2
importlib-metadata==4.8.1
isodate==0.6.0
itsdangerous==1.1.0
jaraco.context==4.0.0
jaraco.docker==2.0
jaraco.functools==3.3.0
Jinja2==2.11.3
joblib==1.0.1
jsmin==2.2.2
livereload==2.6.3
login==0.0.6
lunr==0.6.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
mkdocs-material==7.2.6
mkdocs-material-extensions==1.0.1
mkdocs-minify-plugin==0.4.0
more-itertools==8.8.0
msal==1.14.0
msal-extensions==0.3.0
msrest==0.6.21
nltk==3.6.2
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
opencensus-ext-flask==0.7.5
packaging==21.0
Pillow==8.3.2
pip-upgrader==1.4.15
portalocker==1.7.1
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pymdown-extensions==8.2
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyyaml-env-tag==0.1
regex==2021.8.28
reportlab==3.6.1
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
supervisor==4.2.2
terminaltables==3.1.0
tldextract==3.1.2
tornado==6.1
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
visitor==0.1.3
watchdog==2.1.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.5.0
